+++ docs/conf.py
@@ -38,13 +38,12 @@ html_theme_options = {
     'display_version': True,
     'prev_next_buttons_location': 'bottom',
     'style_external_links': False,
-    'style_nav_header_background': 'white',
+    # 'style_nav_header_background': 'white',
     # Toc options
-    'collapse_navigation': True,
-    'sticky_navigation': True,
-    'navigation_depth': 4,
-    'includehidden': True,
-    'titles_only': False
+    # 'collapse_navigation': False,
+    # 'sticky_navigation': False,
+    # 'navigation_depth': 4,
+    # 'includehidden': True,
 }
 
 html_theme = "sphinx_rtd_theme"