+++ mvsim_tutorial/demo_depth_camera.launch.py
@@ -12,12 +12,14 @@ import os
 
 def generate_launch_description():
     mvsimDir = get_package_share_directory("mvsim")
-    #print('mvsimDir: ' + mvsimDir)
+    # print('mvsimDir: ' + mvsimDir)
 
     # args that can be set from the command line or a default will be used
     world_file_launch_arg = DeclareLaunchArgument(
         "world_file", default_value=TextSubstitution(
             text=os.path.join(mvsimDir, 'mvsim_tutorial', 'demo_depth_camera.world.xml')))
+    do_fake_localization_arg = DeclareLaunchArgument(
+        "do_fake_localization", default_value='True', description='publish tf odom -> base_link')
 
     mvsim_node = Node(
         package='mvsim',
@@ -29,6 +31,7 @@ def generate_launch_description():
                          'mvsim_ros2_params.yaml'),
             {
                 "world_file": LaunchConfiguration('world_file'),
+                "do_fake_localization": LaunchConfiguration('do_fake_localization'),
             }]
     )
 
@@ -42,6 +45,7 @@ def generate_launch_description():
 
     return LaunchDescription([
         world_file_launch_arg,
+        do_fake_localization_arg,
         mvsim_node,
         rviz2_node
     ])