+++ mvsim_tutorial/demo_turtlebot_world.launch.py
@@ -24,6 +24,9 @@ def generate_launch_description():
         "world_file", default_value=TextSubstitution(
             text=MVSIM_WORLD_FILE))
 
+    do_fake_localization_arg = DeclareLaunchArgument(
+        "do_fake_localization", default_value='True', description='publish tf odom -> base_link')
+
     mvsim_node = Node(
         package='mvsim',
         executable='mvsim_node',
@@ -34,6 +37,7 @@ def generate_launch_description():
             {
                 "world_file": LaunchConfiguration('world_file'),
                 "headless": False
+                "do_fake_localization": LaunchConfiguration('do_fake_localization'),
             }]
     )
 
@@ -47,6 +51,7 @@ def generate_launch_description():
 
     return LaunchDescription([
         world_file_launch_arg,
+        do_fake_localization_arg,
         mvsim_node,
         rviz2_node
     ])